+++ dwertheimer.TaskAutomations/changelog.md
@@ -5,14 +5,19 @@
 
 ## What's Changed in this Plugin?
 
+## [2.11.1] (@dwertheimer) 2023-03-12
+
+- Removed error noise on task sorting when lines were note tasks
+
 ## [2.11.0] (@dwertheimer)
 
 ### Added
+
 - Task Sorting: Skip Done/Cancelled; Include Checklist in Sorting
 - React Overdue task processing view v1
 - Overdue: change to checklist type
 
-## [2.10.0] (@dwertheimer) 
+## [2.10.0] (@dwertheimer)
 
 - Added weekly note review question after overdue (and \n'/Review/Reschedule Tasks Scheduled for this week' command